requirements.txt
Flask>=3.0.0
google-cloud-firestore>=2.14.0
gunicorn>=21.2.0
Authlib>=1.3.0
Flask-Login>=0.6.3
python-dotenv>=1.0.0
requests>=2.31.0
